#39 fixed pr comments
